requirements.txt
fastapi==0.115.12
mysql-connector-python==9.3.0 
uvicorn==0.34.2
python-dotenv
groq
langchain_groq
langchain_core
langchain
requests
openai==1.12.0